[#103991306] Add declaration status to email list
See: https://www.pivotaltracker.com/story/show/103991306

This change adds an option to the generate_user_email_list script to
add the declaration status to the CSV output. This moves the logic into
a package in the main app package so that it can be more easily included
into an IPython session for manual testing / debugging. In theory this
makes it easier to write tests for this code although I haven't done for
this script. :/
