+++ app/scripts/__init__.py
@@ -1,0 +1,2 @@
+"""Utilities for scripts
+"""
